CREModel/requirements.txt
grpcio==1.71
grpcio-tools==1.71
protobuf==5.29.3
transformers==4.51.3
numpy==2.2.4
loguru==0.7.3
prometheus-client==0.21.1